photo picker for reports - working on #91

+++ lib/eportfolio_functions.php
@@ -1302,11 +1302,15 @@ function delete_file($filedata){
 	else{
 		mysql_query("DELETE FROM file WHERE id='$file_id';");
 		}
-	
-	if(unlink($filedata['path'])){
-		$success=true;
+
+	$flocation=$filedata['flocation'];
+	$d_f=mysql_query("SELECT * FROM file WHERE location='$flocation';");
+	if(mysql_num_rows($d_f)==1){
+		if(unlink($filedata['path'])){
+			$success=true;
+			}
+		else{trigger_error('Could not remove file from eportfolio: '.$filedata['path'],E_USER_WARNING);}
 		}
-	else{trigger_error('Could not remove file from eportfolio: '.$filedata['path'],E_USER_WARNING);}
  
 
 	return $success;
